separate network setup from shell utilities
